refactor: split treasury budget deficit clamp spec AC cases (#4310)
Extract AC4/AC8 spec case bodies from the 449-line deficit-clamp
module into a sibling cases file so the top 400–449 band gains
headroom under Slice D remainder densify.

Refs #4310
